Adds dropwizard metrics support, for both server/client
