+++ modules/dropwizard/client/src/test/scala/MonitorClientInterceptorTests.scala
@@ -1,0 +1,56 @@
+/*
+ * Copyright 2017-2018 47 Degrees, LLC. <http://www.47deg.com>
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package freestyle.rpc
+package dropwizard
+package client
+
+import com.codahale.metrics.MetricRegistry
+import freestyle.rpc.prometheus.client.{BaseMonitorClientInterceptorTests, ClientRuntime}
+import freestyle.rpc.prometheus.shared.Configuration
+import io.prometheus.client.dropwizard.DropwizardExports
+
+class MonitorClientInterceptorTests extends BaseMonitorClientInterceptorTests {
+
+  override def name: String = "Dropwizard"
+
+  def defaultClientRuntime: ClientRuntime = {
+
+    val metrics: MetricRegistry      = new MetricRegistry
+    val configuration: Configuration = Configuration.defaultBasicMetrics
+    configuration.collectorRegistry.register(new DropwizardExports(metrics))
+
+    ClientRuntime(configuration)
+  }
+
+  def allMetricsClientRuntime: ClientRuntime = {
+
+    val metrics: MetricRegistry      = new MetricRegistry
+    val configuration: Configuration = Configuration.defaultAllMetrics
+    configuration.collectorRegistry.register(new DropwizardExports(metrics))
+
+    ClientRuntime(configuration)
+  }
+
+  def clientRuntimeWithNonDefaultBuckets(buckets: Vector[Double]): ClientRuntime = {
+
+    val metrics: MetricRegistry      = new MetricRegistry
+    val configuration: Configuration = Configuration.defaultAllMetrics.withLatencyBuckets(buckets)
+    configuration.collectorRegistry.register(new DropwizardExports(metrics))
+
+    ClientRuntime(configuration)
+  }
+}
