--- modules/prometheus/server/src/test/scala/RegistryHelper.scala
@@ -1,54 +1,0 @@
-/*
- * Copyright 2017-2018 47 Degrees, LLC. <http://www.47deg.com>
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package freestyle.rpc
-package prometheus
-package server
-
-import io.prometheus.client._
-import scala.collection.JavaConverters._
-
-object RegistryHelper {
-
-  def findRecordedMetric(name: String)(
-      implicit CR: CollectorRegistry): Option[Collector.MetricFamilySamples] =
-    CR.metricFamilySamples.asScala.find(_.name == name)
-
-  def findRecordedMetricOrThrow(name: String)(
-      implicit CR: CollectorRegistry): Collector.MetricFamilySamples =
-    findRecordedMetric(name).getOrElse(
-      throw new IllegalArgumentException(s"Could not find metric with name: $name"))
-
-  def extractMetricValue(name: String)(implicit CR: CollectorRegistry): Double = {
-    val result = findRecordedMetricOrThrow(name)
-    result.samples.asScala.headOption
-      .map(_.value)
-      .getOrElse(throw new IllegalArgumentException(
-        s"Expected one value, but got ${result.samples.size} for metric $name"))
-  }
-
-  def countSamples(metricName: String, sampleName: String)(implicit CR: CollectorRegistry): Int =
-    CR.metricFamilySamples.asScala
-      .filter(_.name == metricName)
-      .map { sample =>
-        sample.samples.asScala.count(_.name == sampleName)
-      }
-      .sum
-
-  def printRegistry(collectorRegistry: CollectorRegistry): Unit =
-    collectorRegistry.metricFamilySamples.asScala.foreach(println)
-
-}
